feat: Add boxed precompile trait
