+++ crates/primitives/Cargo.toml
@@ -33,6 +33,7 @@ once_cell = { version = "1.19", default-features = false, optional = true }
 enumn = "0.1"
 derive_more = { version = "0.99", optional = true }
 cfg-if = "1"
+dyn-clone = "1.0"
 
 # optional
 serde = { version = "1.0", default-features = false, features = [
